Two files pertinent to comparing the data.
These files are needed to compare the data as they along with the gettingosmdatacopy file would give the number of gtfs stops, osm stops that are bus stops and osm stops which contain certain latitudes. The intention is to have the gettingosmdatacopy file compare stops based on latitude and longitude and compare their distances using the vincenty script.